requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-filter==24.2
django-scheduler==0.10.1
djangorestframework==3.15.1
environs==8.0.0
frozenlist==1.4.1
icalendar==5.0.12
idna==3.6
magic-filter==1.0.12
marshmallow==3.21.0
multidict==6.0.5
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
yarl==1.9.4
